♻️ refactor: update imports, prop type names and names of css modules
