+++ src/app/App.tsx
@@ -1,9 +1,9 @@
-import Header from "../widgets/LayoutHeader";
-import Footer from "../widgets/LayoutFooter";
-import PostList from "../widgets/PostList";
-import MainLayout from "../shared/layouts/MainLayout";
+import { Header } from "../widgets/LayoutHeader";
+import { Footer } from "../widgets/LayoutFooter";
+import { PostList } from "../widgets/PostList";
+import { MainLayout } from "../shared/layouts/MainLayout";
 
-export default function App() {
+export function App() {
   return (
     <MainLayout header={<Header />} footer={<Footer />}>
       <PostList />
