+++ src/widgets/LayoutFooter/index.tsx
@@ -1,6 +1,6 @@
-import styles from "./index.module.css";
+import styles from "./Footer.module.css";
 
-export default function Footer() {
+export function Footer() {
   return (
     <footer className={styles.footer}>
       <div className={styles.author}>Created By Elagin Anton</div>
