+++ src/widgets/LayoutHeader/index.tsx
@@ -1,6 +1,6 @@
-import styles from "./index.module.css";
+import styles from "./Header.module.css";
 
-export default function Header() {
+export function Header() {
   return (
     <header className={styles.header}>
       <h1 className={styles.title}>Post List</h1>
